move the C compression types from server.rs into gateway.rs
They were only ever consumed by the gateway, and living in the ungated
websocket-server file forced six item-level remote-access cfg gates —
the only cfg attributes in that file. In gateway.rs the module-level
gate in lib.rs covers them, matching the crate convention. A crate-root
re-export preserves the public path that pub use server::* provided
(without it, the FFI-constructed enum variants trip the dead-code
lint). The regenerated header keeps every declaration in place but
simplifies their guards to defined(FOXGLOVE_REMOTE_ACCESS), matching
every other gateway type; the compression types were the only ones
still carrying the !__wasm__ term from server.rs's module gate.
